[#122] refactor: consultation command 제거
