+++ src/main/java/org/silsagusi/joonggaemoa/domain/consultation/service/dto/ConsultationMonthInformResponse.java
@@ -1,17 +1,18 @@
-package org.silsagusi.joonggaemoa.domain.consultation.service.command;
+package org.silsagusi.joonggaemoa.domain.consultation.service.dto;
 
 import lombok.Builder;
 import lombok.Getter;
 
 import java.util.List;
 
-@Builder
 @Getter
-public class ConsultationMonthInformCommand {
+@Builder
+public class ConsultationMonthInformResponse {
     private Long consultationAll;
     private Long consultationWaiting;
     private Long consultationConfirmed;
     private Long consultationCancelled;
     private Long consultationCompleted;
     private List<Integer> daysCount;
+
 }
