+++ src/main/java/org/silsagusi/joonggaemoa/domain/consultation/service/dto/ConsultationSummaryResponse.java
@@ -1,0 +1,11 @@
+package org.silsagusi.joonggaemoa.domain.consultation.service.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ConsultationSummaryResponse {
+    private Integer todayCount;
+    private Integer remainingCount;
+}
