+++ src/main/java/org/silsagusi/joonggaemoa/domain/consultation/service/dto/UpdateConsultationRequest.java
@@ -1,0 +1,18 @@
+package org.silsagusi.joonggaemoa.domain.consultation.service.dto;
+
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class UpdateConsultationRequest {
+    private LocalDateTime date;
+    private String purpose;
+    private String interestProperty;
+    private String interestLocation;
+    private String contractType;
+    private String assetStatus;
+    private String memo;
+    private String consultationStatus;
+
+}
